Elasticsearch: Rename public/app/plugins/datasource/elasticsearch/elastic_response.ts to follow convention
