+++ public/app/plugins/datasource/elasticsearch/ElasticResponse.test.ts
@@ -1,9 +1,10 @@
 import { DataFrameView, FieldCache, KeyValue, MutableDataFrame } from '@grafana/data';
 import flatten from 'app/core/utils/flatten';
 
-import { ElasticResponse } from '../elastic_response';
-import { highlightTags } from '../query_def';
 import { ElasticsearchQuery } from '../types';
+
+import { ElasticResponse } from './ElasticResponse';
+import { highlightTags } from './query_def';
 
 describe('ElasticResponse', () => {
   let targets: ElasticsearchQuery[];
