+++ public/app/plugins/datasource/elasticsearch/datasource.ts
@@ -33,6 +33,7 @@ import { getTemplateSrv, TemplateSrv } from 'app/features/templating/template_sr
 
 import { RowContextOptions } from '../../../features/logs/components/LogRowContextProvider';
 
+import { ElasticResponse } from './ElasticResponse';
 import LanguageProvider from './LanguageProvider';
 import { ElasticsearchAnnotationsQueryEditor } from './components/QueryEditor/AnnotationQueryEditor';
 import {
@@ -46,7 +47,6 @@ import {
   Logs,
 } from './components/QueryEditor/MetricAggregationsEditor/aggregations';
 import { metricAggregationConfig } from './components/QueryEditor/MetricAggregationsEditor/utils';
-import { ElasticResponse } from './elastic_response';
 import { IndexPattern } from './index_pattern';
 import { ElasticQueryBuilder } from './query_builder';
 import { defaultBucketAgg, hasMetricOfType } from './query_def';
